chore: shards be optional
License: MIT
