+++ packages/upload-client/test/helpers/random.js
@@ -1,8 +1,4 @@
-import { CarWriter } from '@ipld/car'
-import { CID } from 'multiformats/cid'
-import * as raw from 'multiformats/codecs/raw'
-import { sha256 } from 'multiformats/hashes/sha2'
-import * as CAR from '@ucanto/transport/car'
+import { toCAR } from './car.js'
 
 /** @param {number} size */
 export async function randomBytes(size) {
@@ -31,19 +27,5 @@ export async function randomBytes(size) {
 /** @param {number} size */
 export async function randomCAR(size) {
   const bytes = await randomBytes(size)
-  const hash = await sha256.digest(bytes)
-  const root = CID.create(1, raw.code, hash)
-
-  const { writer, out } = CarWriter.create(root)
-  writer.put({ cid: root, bytes })
-  writer.close()
-
-  const chunks = []
-  for await (const chunk of out) {
-    chunks.push(chunk)
-  }
-  const blob = new Blob(chunks)
-  const cid = await CAR.codec.link(new Uint8Array(await blob.arrayBuffer()))
-
-  return Object.assign(blob, { cid, roots: [root] })
+  return toCAR(bytes)
 }
